requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1;python_version<"3.9"
bibtexparser==1.3.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==38.0.1
distinctipy==1.2.3
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.3
django-allauth==0.51.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-environ==0.11.2
django-filter==21.1
django_geojson==4.0.0
django-heroku==0.3.1
django-mathfilters==1.0.0
djangorestframework==3.13.1
djangorestframework-gis==1.0
django-recaptcha==4.0.0
feedparser==6.0.10
geopandas==0.13.2
gunicorn==20.1.0
idna==3.3
matplotlib==3.7.4
MarkupSafe==2.1.1
oauthlib==3.2.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
python-decouple==3.6
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
pyzotero==1.5.5
requests==2.27.1
requests-oauthlib==1.3.1
sentry-sdk==1.5.10
sgmllib3k==1.0.0
Shapely==1.8.2
six==1.16.0
sqlparse==0.4.2
types-cryptography==3.3.23
urllib3==1.26.9
whitenoise==6.0.0
